+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [2.0] - 2024-07-24
+### Added
+- `build.yml`, `deploy.yml`: converted project to GitHub Actions
+### Changed
+- `RefResourceLoader`: added `addToCache` function
+- `ResourceRef`: added comments
+- `pom.xml`: updated Kotlin version to 1.9.24
+### Removed
+- `.travis.yml`
+
 ## [1.5] - 2024-02-25
 ### Changed
 - `pom.xml`: updated dependency version